commands/requirements.txt
discord.py
aiosqlite
requests
feedparser
beautifulsoup4
yt-dlp
pytz
